feat(market-data): add watchlist event handling for data ingestion and update api endpoints
- Change API route for `CreateAssetEndpoint` to `/api/v1/assets` and add authorization and response type configurations.
- Add `IMessageBus` dependency to `CreateAssetHandler` for publishing `AssetWatchlistChangedEvent`.
- Modify `AssetWatchlistChangedEvent` to use strings for `Action` and `AssetType` properties.
- Rename `SaveMarketDataCommand` namespace for clarity.
- Change `Volume` property type from `long` to `decimal` in `SaveMarketDataHandler`.
- Enhance `BinanceDataIngestionService` with methods for managing asset watchlist changes.
- Introduce `CombinedStreamModel<T>` for better deserialization of Binance WebSocket messages.
- Update project structure to include a new folder for quotes infrastructure features.
- Modify `appsettings.Development.json` to correct the Binance WebSocket URI.

+++ src/contexts/market-data/src/FinnHub.MarketData.WebApi/Features/Quotes/Domain/Entities/HistoricalQuote.cs
@@ -10,7 +10,7 @@ public sealed class HistoricalQuote : Entity
     public decimal High { get; private set; }
     public decimal Low { get; private set; }
     public decimal Close { get; private set; }
-    public long Volume { get; private set; }
+    public decimal Volume { get; private set; }
     public string Interval { get; private set; }
     public DateTime CreatedAt { get; private set; } = DateTime.UtcNow;
 
@@ -21,7 +21,7 @@ public sealed class HistoricalQuote : Entity
         decimal high,
         decimal low,
         decimal close,
-        long volume,
+        decimal volume,
         string interval
     )
     {
